pyproject.toml
[tool.poetry]
name = "dndata-server"
version = "0.1.0"
description = ""
authors = ["Michal Frystacky <codefry@pm.me>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.2"
graphene-django = "^2.13.0"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
